Refactor `SaveTabHistoryEntryUseCase`
Currently, the return from this use case is the id of the stored `HistoryPage`, which may not be intuitive as it requires callers to manage the retrieval of the (stored) `HistoryPage` by id.

Return the updated or stored `HistoryPage`, otherwise returns null. This change eliminates the extra step to retrieve a `HistoryPage` and make the result more useful.